microservices/pyproject.toml
[project]
name = "microservices"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "assemblyai>=0.48.4",
    "fastapi[standard]>=0.128.0",
    "langchain>=1.2.7",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.0",
    "langgraph>=1.0.7",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "werkzeug>=3.1.5",
]
